funcionalidades completadas

+++ src/services/productoServices.ts
@@ -12,7 +12,7 @@ export const getAllProductos = async (): Promise<Producto[]> => {
 };
 
 // Agregar un nuevo producto
-export const addProducto = async (productoData: Partial<Producto> & { precio: number }): Promise<Producto | null> => {
+export const addProducto = async (productoData: Partial<Producto> & { precio: number }): Promise<Producto_Precio[]| null> => {
     if ((await getProductosByName(String(productoData.nombre))).length == 0) {
         const productoRepository = AppDataSource.getRepository(Producto);
         const precioRepository = AppDataSource.getRepository(Producto_Precio);
@@ -24,7 +24,15 @@ export const addProducto = async (productoData: Partial<Producto> & { precio: nu
         // Crear un nuevo precio asociado al producto
         const newPrecio = precioRepository.create({ precio: productoData.precio, producto: newProducto });
         await precioRepository.save(newPrecio);
-        return newProducto; // Devolver el producto recién creado
+
+        // Agregar el precio recién creado al producto
+        if(!newProducto.producto_precios) {
+            newProducto.producto_precios = [newPrecio];
+        } else {
+            newProducto.producto_precios.push(newPrecio);
+        }
+        
+        return newProducto.producto_precios; // Devolver el producto recién creado
     }
     return null;
 };
@@ -46,16 +54,25 @@ export const getProductoByName = async (name: string): Promise<Producto | null> 
 
 // Actualizar un producto
 export const updateProducto = async (id: number, productoData: Partial<Producto>): Promise<Producto | null> => {
-    if((await getProductosByName(String(productoData.nombre))).length<=1 ){
     const productoRepository = AppDataSource.getRepository(Producto);
-    await productoRepository.update(id, productoData);
-    return await productoRepository.findOneBy({ id_producto: id });
-    }else
+    const element = await productoRepository.findOneBy({ id_producto: id });
+    //si existe el producto
+    if(element){
+        const elementWithSameName = await getProductoByName(String(productoData.nombre))
+        //si el nombre pasado es diferente al nombre existente y no existe otro elemento con ese nombre pasado como parámetro
+        if((element.nombre != productoData.nombre && !elementWithSameName) || element.nombre == productoData.nombre){
+            const productoRepository = AppDataSource.getRepository(Producto);
+            await productoRepository.update(id, productoData);
+            //devolver elemento actualizado
+            return await productoRepository.findOneBy({ id_producto: id });
+        }
+
+    }
     return null;
 };
 
-//mover producto a una tienda
-export const moveProducto = async (idTiendaProductoPrecio: number, idTienda:number): Promise<TiendaProductoPrecio | null> => {
+//editar tienda en tiendaProductoPrecio
+export const moveTiendaProductoPrecio = async (idTiendaProductoPrecio: number, idTienda:number): Promise<TiendaProductoPrecio | null> => {
     const tiendaProductoPrecioRepository = AppDataSource.getRepository(TiendaProductoPrecio);
     const tiendaRepository = AppDataSource.getRepository(Tienda);
 
@@ -82,36 +99,72 @@ export const moveProducto = async (idTiendaProductoPrecio: number, idTienda:numb
     return elemento; // Devolver el registro actualizado
 };
 
-// Eliminar un producto
+// Mover producto y precios a nueva tienda
+export const moveProducto = async (idProducto: number, idNuevaTienda: number): Promise<TiendaProductoPrecio[] | null> => {
+    const productoRepository = AppDataSource.getRepository(Producto);
+    const tiendaProductoPrecioRepository = AppDataSource.getRepository(TiendaProductoPrecio);
+    
+    // Buscar el producto por su ID y obtener todos los Producto_Precio asociados
+    const producto = await productoRepository.findOne({ 
+        where: { id_producto: idProducto },
+        relations: ['producto_precios'] 
+    });
+
+    if (!producto) {
+        return null; // Devolver null si no se encuentra el producto
+    }
+
+    const productosPrecio = producto.producto_precios;
+
+    const tiendaProductoPrecioActualizados: TiendaProductoPrecio[] = [];
+
+    for (const productoPrecio of productosPrecio) {
+        // Obtener todos los registros de TiendaProductoPrecio asociados al Producto_Precio
+        const tiendaProductoPrecios = await tiendaProductoPrecioRepository.find({ 
+            where: { producto_precio: { id_producto_precio: productoPrecio.id_producto_precio } } 
+        });
+
+        for (const tiendaProductoPrecio of tiendaProductoPrecios) {
+            // Utilizar el servicio existente para cambiar la tienda en TiendaProductoPrecio
+            const tiendaProductoActualizado = await moveTiendaProductoPrecio(tiendaProductoPrecio.id_tiendaProductoPrecio, idNuevaTienda);
+            
+            if (tiendaProductoActualizado) {
+                tiendaProductoPrecioActualizados.push(tiendaProductoActualizado);
+            }
+        }
+    }
+
+    return tiendaProductoPrecioActualizados;
+};
+
+
+// Eliminar un producto y sus registros relacionados en las entidades vinculadas
 export const deleteProducto = async (id: number): Promise<boolean> => {
     const productoRepository = AppDataSource.getRepository(Producto);
+    const productoPrecioRepository = AppDataSource.getRepository(Producto_Precio);
 
-    // Obtener el producto por su ID
-    const producto = await productoRepository.findOneBy({id_producto: id});
+    // Obtener el producto por su ID con todas las relaciones cargadas
+    const producto = await productoRepository.findOneBy({id_producto : id});
 
     if (!producto) {
         return false; // Devolver false si no se encuentra el producto
     }
 
-    // Eliminar los registros relacionados en la tabla Producto_Precio
-    await producto.producto_precios.forEach(async (productoPrecio) => {
-        await AppDataSource.getRepository(Producto_Precio).delete(productoPrecio.id_producto_precio);
-    });
+    const precio = await productoPrecioRepository.findOneBy({producto : producto});
+    
+    if (!precio){
+          // Eliminar el producto en la tabla Producto
+         const result = await productoRepository.delete(id);
 
-    // Eliminar el producto en la tabla Producto
-    const result = await productoRepository.delete(id);
-
-    // Verificar que se haya eliminado al menos un registro
-    return result.affected !== null && result.affected !== undefined && result.affected > 0;
+         // Verificar que se haya eliminado al menos un registro
+         return result.affected !== null && result.affected !== undefined && result.affected > 0;
+    }else
+    return false;//el producto tiene un precio en Producto_Precio, por lo que no se puede eliminar
 };
 
 
 // Obtener productos por nombre (array)
 export const getProductosByName = async (name: string): Promise<Producto[]> => {
-    if (!name) {
-        throw new Error("El nombre no puede ser vacío");
-    }
-    
     const productoRepository = AppDataSource.getRepository(Producto);
     const normalizedName = removeAccents(name.toLowerCase()); // Normalizar el nombre buscado
